+++ src/components/Elements/Tab/Tab.js
@@ -3,31 +3,22 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import TabContent from './TabContent';
 import TabButton from './TabButton';
-import { BaseProps, Color } from '../../../shared';
+import { BaseProps, Color, warn } from '../../../shared';
 import './tab.scss';
 
 /**
  * The filterTabChildren function filters out all elements different from TabContent type.
- * It returns an object which contain the cloned element plus some props usefull to create tab buttons.
+ * It returns an object which contain the cloned element plus some props useful to create tab buttons.
  */
 const filterTabChildren = (child, index, props) => {
   if (child.type !== TabContent) {
-    /**
- * Eslint forces the developer to not have any `console` statement, in this case we want to warn the
- * user without throwing an error so it's perfectly safe to disable this rule.
- */
-    /* eslint-disable-next-line no-console */
-    console.warn(
-      'Tab allows only Tab.Content children, other kind of elements will be wiped out',
-    );
+    warn('Tab allows only Tab.Content children, other kind of elements will be wiped out');
     return null;
   }
 
   const result = Object.create(null);
 
-  result.child = React.cloneElement(child, {
-    active: props.active === index,
-  });
+  result.child = React.cloneElement(child, { active: props.active === index });
   result.title = child.props.title;
   result.icon = child.props.icon;
   result.disabled = child.props.disabled;
@@ -37,7 +28,7 @@ const filterTabChildren = (child, index, props) => {
 
 
 /**
- * The Tab component consists of clickable labels that shows the corrisponding content.
+ * The Tab component consists of clickable labels that shows the corresponding content.
  */
 // the React.memo has been used here rather than on the export line like other cases, to avoid wrapping the shortcut.
 const Tab = React.memo((props) => {
@@ -79,18 +70,18 @@ Tab.propTypes = {
    */
   color: Color,
   /**
-  * The callback to be performed on content change
-  */
+   * The callback to be performed on content change
+   */
   onChange: PropTypes.func.isRequired,
   /**
-  * Defines the current active tab index
-  */
+   * Defines the current active tab index
+   */
   active: PropTypes.oneOfType([
     PropTypes.number,
     PropTypes.string,
   ]).isRequired,
   /**
-  /**
+   /**
    * This props defines the orientation of the tabs.
    */
   orientation: PropTypes.string,