+++ src/shared/index.js
@@ -4,6 +4,7 @@ export { default as Color } from './types/Color';
 export { default as Size } from './types/Size';
 export { default as BaseFormProps } from './types/BaseFormProps';
 export { default as IconProp } from './types/IconProp';
+export { default as PillProp } from './types/PillProp';
 
 // functions
 export { default as makeCallback } from './functions/makeCallback';
@@ -13,3 +14,4 @@ export { default as makePillFromProp } from './functions/makePillFromProp';
 export { default as emptyChildren } from './functions/emptyChildren';
 export { default as getLastChild } from './functions/getLastChild';
 export { default as getFirstChild } from './functions/getFirstChild';
+export { default as warn } from './functions/warn';